src/python/requirements.txt
boto3
cloudpickle
google-cloud-storage
numpy
pandas
pydantic
SQLAlchemy
distro
pyyaml
typing_extensions>=4.3.0
Pillow
packaging
pymongo
aqueduct-sdk==0.1.5
